feat: add new root navigation + layout
