+++ packages/apps/spaces/.eslintrc.json
@@ -31,7 +31,12 @@
     "@typescript-eslint/prefer-ts-expect-error": "warn",
     "prettier/prettier": "warn",
     "no-console": "warn",
-    "react/display-name": "off"
+    "react/display-name": "off",
+    "react-hooks/exhaustive-deps": "off",
+    "@typescript-eslint/no-unused-vars": [
+      "warn",
+      { "varsIgnorePattern": "^_", "ignoreRestSiblings": true, "args": "none" }
+    ]
   },
   "ignorePatterns": ["**/*.generated.ts"],
   "overrides": [
