--- packages/apps/spaces/app/components/Sidebar/LogoutSidebarItem.tsx
@@ -1,24 +1,0 @@
-'use client';
-import { signOut } from 'next-auth/react';
-
-import Exit from '@spaces/atoms/icons/Exit';
-import { useJune } from '@spaces/hooks/useJune';
-
-import { SidebarItem } from './SidebarItem';
-
-export const LogoutSidebarItem = () => {
-  const analytics = useJune();
-
-  const handleClick = () => {
-    analytics?.reset();
-    signOut();
-  };
-
-  return (
-    <SidebarItem
-      label='Logout'
-      onClick={handleClick}
-      icon={<Exit height={24} width={24} style={{ scale: '0.8' }} />}
-    />
-  );
-};
