+++ packages/apps/spaces/app/organizations/[id]/components/MainSection/MainSection.tsx
@@ -6,10 +6,11 @@ export const MainSection = ({ children }: { children?: React.ReactNode }) => {
   return (
     <Flex
       flex='3'
-      h='calc(100vh - 2rem)'
+      h='calc(100vh - 1rem)'
       bg='white'
       borderRadius='2xl'
-      shadow='base'
+      border='1px solid'
+      borderColor='gray.200'
     >
       {children}
     </Flex>
