+++ packages/apps/spaces/app/organizations/[id]/components/SideSection/SIdeSection.tsx
@@ -4,7 +4,7 @@ import { Flex } from '@ui/layout/Flex';
 
 export const SideSection = ({ children }: { children?: React.ReactNode }) => {
   return (
-    <Flex flex='1' h='calc(100vh - 2rem)' minW='28rem'>
+    <Flex flex='1' h='calc(100vh - 1rem)' minW='28rem'>
       {children}
     </Flex>
   );
