+++ packages/apps/spaces/components/organization/organization-list/OrganizationListColumns.tsx
@@ -1,5 +1,4 @@
-import { createColumnHelper } from '@spaces/ui/presentation/table/Table';
-import { THead } from '@ui/presentation/_Table';
+import { THead, createColumnHelper } from '@spaces/ui/presentation/Table';
 import {
   AddressTableCell,
   OrganizationTableCell,
@@ -30,7 +29,6 @@ export const columns = [
         />
       );
     },
-    minSize: 200,
     header: (props) => (
       <THead<Organization>
         title='Company'
@@ -71,7 +69,6 @@ export const columns = [
         </>
       );
     },
-    minSize: 200,
     skeleton: () => (
       <div className={styles.skeletonWrapper}>
         <Skeleton width='100%' height='21px' />
@@ -86,7 +83,6 @@ export const columns = [
       if (!url) return <div />;
       return <ExternalLinkCell url={url} />;
     },
-    minSize: 200,
     header: (props) => (
       <THead<Organization> title='Domain' subTitle='Website' {...props} />
     ),
@@ -97,7 +93,6 @@ export const columns = [
     cell: (props) => {
       return <AddressTableCell locations={props.getValue()} />;
     },
-    minSize: 200,
     header: (props) => (
       <THead<Organization> title='Location' subTitle='Address' {...props} />
     ),
@@ -111,7 +106,6 @@ export const columns = [
         organizationId={props.row.original.id}
       />
     ),
-    minSize: 200,
     header: (props) => <THead<Organization> title='Owner' {...props} />,
     skeleton: () => <Skeleton width='100%' height='21px' />,
   }),
@@ -123,7 +117,6 @@ export const columns = [
         healthIndicator={props.row.original.healthIndicator}
       />
     ),
-    minSize: 200,
     header: (props) => <THead<Organization> title='Health' {...props} />,
     skeleton: () => <Skeleton width='100%' height='21px' />,
   }),
@@ -138,7 +131,6 @@ export const columns = [
         }
       />
     ),
-    minSize: 200,
     header: (props) => (
       <THead<Organization>
         title='Last touchpoint'
