--- packages/apps/spaces/ui/presentation/table/TActions.module.scss
@@ -1,12 +1,0 @@
-@import 'styles/mixins';
-
-.actionsMenuButton {
-  width: 30px !important;
-  height: 30px;
-}
-
-.actionHeader {
-  height: 100%;
-  @include justify-end;
-  align-items: center;
-}
