--- packages/apps/spaces/ui/presentation/table/TActions.tsx
@@ -1,51 +1,0 @@
-import { forwardRef, useRef } from 'react';
-import type { MenuProps } from 'primereact/menu';
-
-import EllipsesV from '@spaces/atoms/icons/EllipsesV';
-import { IconButton } from '@spaces/atoms/icon-button/IconButton';
-import { OverlayPanel } from '@spaces/atoms/overlay-panel/OverlayPanel';
-
-import styles from './TActions.module.scss';
-
-interface TActionsProps {
-  model: MenuProps['model'];
-}
-
-export const TActions = forwardRef<HTMLDivElement, TActionsProps>(
-  ({ model }, ref) => {
-    const op = useRef(null);
-
-    return (
-      <div className={styles.actionHeader} ref={ref}>
-        <IconButton
-          label='Actions'
-          className={styles.actionsMenuButton}
-          id={'finder-actions-dropdown-button'}
-          mode='secondary'
-          size='xxxxs'
-          //@ts-expect-error revisit
-          onClick={(e) => op?.current?.toggle(e)}
-          icon={
-            <EllipsesV
-              height={24}
-              width={24}
-              style={{ transform: 'rotate(90deg)' }}
-            />
-          }
-        />
-
-        <OverlayPanel
-          ref={op}
-          model={model}
-          style={{
-            maxHeight: '400px',
-            height: 'fit-content',
-            overflowX: 'hidden',
-            overflowY: 'auto',
-            bottom: 0,
-          }}
-        />
-      </div>
-    );
-  },
-);
