--- packages/apps/spaces/ui/presentation/table/THead.module.scss
@@ -1,26 +1,0 @@
-.thead {
-  display: flex;
-  width: 100%;
-  flex-direction: column;
-  align-items: flex-start;
-  justify-content: flex-start;
-  font-weight: normal;
-
-  &.labelWithAvatar {
-    margin-left: calc(
-      30px + 1rem
-    ); // todo change after adding other column avatar length + margin
-  }
-
-  &.withIcon {
-    margin-left: 24px;
-    padding-left: 8px;
-  }
-}
-.title {
-  font-weight: 500;
-}
-.subTitle {
-  color: var(--table-sublabel-color);
-  font-size: var(--font-size-sm);
-}
