--- packages/apps/spaces/ui/presentation/table/Table.module.scss
@@ -1,188 +1,0 @@
-@import 'styles/mixins';
-
-$width: calc(
-  100vw - 134px - 2.4rem
-); //80px metapanel + 24 grid gap + 24 margin right + 2.4 rem padding
-
-.container {
-  width: $width;
-  display: flex;
-  flex-direction: column;
-}
-
-.table {
-  overflow-x: auto;
-  overflow-y: hidden;
-  border-collapse: collapse;
-  border-style: hidden;
-  border-radius: var(--border-radius-sm);
-  width: $width;
-  height: calc(100vh - 180px); // 100vh - top search bar
-}
-
-.thead {
-  & .tr {
-    background: #f0f0f0;
-    border-top-right-radius: 6px;
-    border-top-left-radius: 6px;
-  }
-}
-
-.tbody {
-  @include scrollbar;
-  overflow-y: auto;
-  overflow-x: hidden;
-  height: inherit;
-  width: $width;
-  position: relative;
-  display: block;
-}
-
-.row {
-  display: flex;
-  position: absolute;
-  top: 0;
-  left: 0;
-  width: 100%;
-  overflow: visible;
-  font-size: var(--font-size-sm);
-  padding-top: var(--spacing-xs);
-  padding-bottom: var(--spacing-xs);
-
-  &.even {
-    background: #f9f9f9;
-  }
-
-  &.emptyRow {
-    display: flex;
-    justify-content: center;
-  }
-}
-
-.rowCell {
-  white-space: nowrap;
-  word-break: keep-all;
-  padding-left: var(--spacing-xs);
-}
-
-.th {
-  display: block;
-  padding: var(--spacing-xs) var(--spacing-xs);
-  white-space: nowrap;
-  word-break: keep-all;
-  font-weight: 500;
-  text-overflow: ellipsis;
-  font-size: var(--font-size-sm);
-  border-top: var(--border-light);
-  border-left: var(--border-light);
-
-  &:first-of-type {
-    border-top-left-radius: var(--border-radius-sm);
-  }
-  &:last-of-type {
-    border-top-right-radius: var(--border-radius-sm);
-    border-right: var(--border-light);
-  }
-  &.selectCell {
-    border-right: 1px solid transparent;
-  }
-  &.actionCell {
-    border-left: 1px solid transparent;
-  }
-}
-
-.tr {
-  display: flex;
-}
-
-.totalItems {
-  font-size: var(--font-size-sm);
-  align-self: flex-end;
-}
-
-.selectCell {
-  width: 24px;
-  padding-left: var(--spacing-xs);
-}
-.selectCheckboxWrapper {
-  height: 24px;
-  display: flex;
-  align-items: center;
-  flex-direction: row;
-}
-
-.selectCheckbox {
-  --active: var(--royal-blue-800);
-  --active-inner: #fff;
-  --focus: 1px var(--royal-blue-900);
-  --border: var(--gray-500);
-  --border-hover: var(--royal-blue-800);
-  --background: #fff;
-  --disabled: var(--gray-300);
-  --disabled-inner: var(--gray-200);
-  -webkit-appearance: none;
-  -moz-appearance: none;
-  border-radius: 2px;
-  outline: none;
-  display: inline-block;
-  vertical-align: top;
-  position: relative;
-  margin: 0;
-  cursor: pointer;
-  border: 1px solid var(--bc, var(--border));
-  background: var(--b, var(--background));
-  transition: background 0.3s, border-color 0.3s, box-shadow 0.2s;
-  &:after {
-    content: '';
-    display: block;
-    left: 0;
-    top: 0;
-    position: absolute;
-    transition: transform var(--d-t, 0.3s) var(--d-t-e, ease),
-      opacity var(--d-o, 0.2s);
-  }
-  &:checked {
-    --b: var(--active);
-    --bc: var(--active);
-    --d-o: 0.3s;
-    --d-t: 0.6s;
-    --d-t-e: cubic-bezier(0.2, 0.85, 0.32, 1.2);
-  }
-  &:disabled {
-    --b: var(--disabled);
-    cursor: not-allowed;
-    opacity: 0.9;
-    &:checked {
-      --b: var(--disabled-inner);
-      --bc: var(--border);
-    }
-    & + span {
-      cursor: not-allowed;
-    }
-  }
-  &:hover {
-    &:not(:checked) {
-      &:not(:disabled) {
-        --bc: var(--border-hover);
-      }
-    }
-  }
-  &:focus {
-    box-shadow: 0 0 0 var(--focus);
-  }
-  &:not(.switch) {
-    &:after {
-      opacity: var(--o, 0);
-    }
-    &:checked {
-      --o: 1;
-    }
-  }
-  & + span {
-    font-size: inherit;
-    display: inline-block;
-    vertical-align: top;
-    cursor: pointer;
-    margin-left: 4px;
-  }
-}
